requirements.txt
# Core dependencies
google-generativeai>=0.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
requests>=2.31.0

# Utilities
pillow>=10.2.0
python-dateutil>=2.8.2
schedule>=1.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=24.0.0
ruff>=0.1.0
